\name{dep}
\alias{dep}
\docType{data}
\title{Example of dependent/explained Variable for Markov switching model}
\description{Just an example data, in this case logarithm returns. This example data is the same one from the MatLab version of the package}
\usage{data(dep)}
\format{The format is:
 num [1:1000] -0.07318  0.05899  0.00795  0.00000  0.00717 ...}

\examples{
data(dep)
## maybe str(dep) ; plot(dep) ...
}
\keyword{datasets}
