\name{dim.MS_Model}
\alias{dim.MS_Model}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Function to Get the dimension/Size of Markov Switching Variables}
\description{This function will output the dimensions of the model, that is, how many observations, how many parameters, how many states, etc..}
\usage{
dim.MS_Model(MS_Model_In)
}

\arguments{
  \item{MS_Model_In}{A S4 Object obtained with MS_Regress_Fit}
}

\value{Returns nothing. Prints the output in R screen}
\author{Marcelo Perlin - ICMA Centre/UK}

\seealso{\code{\link{MS_Regress_Fit}},\code{\link{MS_Regress_Simul}},\code{\link{MS_Regress_For}} }


