\name{fMarkovSwitching-package}
\alias{fMarkovSwitching-package}
\alias{fMarkovSwitching}
\docType{package}
\title{Package for Estimation, Simulation and Forecasting of a Markov Switching Model}
\description{This package has designed the estimation, simulation and forecasting of Markov Switching models. See examples for simple ways of using the package and also check the pdf document for advanced examples of functions}
\details{
\tabular{ll}{
Package: \tab fMarkovSwitching\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2008-10-27\cr
License: \tab GPL (>=2) \cr
LazyLoad: \tab yes\cr
}

}
\author{Marcelo Perlin - ICMA/UK <\email{marceloperlin@gmail.com}>}
\references{
\cite{ALEXANDER, C. (2008) 'Market Risk Analysis: Practical Financial Econometrics' Wiley}\cr
\cite{HAMILTON, J., D. (1994) 'Time Series Analysis' Princeton University Press}\cr
\cite{HAMILTON, J., D. (2005) 'Regime Switching Models' Palgrave Dictionary of Economic}\cr
\cite{KIM, C., J., NELSON, C., R. (1999) 'State Space Model with Regime Switching: Classical and Gibbs-Sampling Approaches with Applications' The MIT press}
}

\keyword{ package }
\examples{#See each function example (or pdf document)..}
