\name{indep}
\alias{indep}
\docType{data}
\title{Example of independent/explanatory Variable for Markov switching model}
\description{Just an example data, in this case logarithm returns. This example data is the same one from the MatLab version of the package}
\usage{data(indep)}
\format{
  The format is:
 num [1:1000, 1:3]
}
\details{The data consists of one vector or ones (the constant) and 2 vectors of time series}

\examples{# See MS_Regress_Fit for example}
\keyword{datasets}
